Pull Request

+++ src/libext/libxc/build_libxc.sh
@@ -7,14 +7,25 @@ check_tgz() {
     [ -f $1 ] && gunzip -t $1 > /dev/null && myexit=1
     echo $myexit
 }
-
-VERSION=5.2.2
+if [ $# -eq 0 ];  then
+    VERSION=6.0.0
+else
+    VERSION=$1
+fi
+VERSION_MAJOR=$(echo $VERSION | cut -d . -f 1)
+if [[ "$VERSION_MAJOR" -lt 4 ]]; then
+    echo
+    echo "LIBXC unsupported version " "$VERSION"
+    echo "please use 4.0.0 and later versions"
+    echo
+    exit  1
+fi
 TGZ=libxc-${VERSION}.tar.gz
 if [ `check_tgz $TGZ` == 1 ]; then
     echo "using existing $TGZ"
 else
     echo "downloading $TGZ"
-    curl -L https://gitlab.com/libxc/libxc/-/archive/${VERSION}/libxc-${VERSION}.tar.gz -o $TGZ
+    curl -sS -L https://gitlab.com/libxc/libxc/-/archive/${VERSION}/libxc-${VERSION}.tar.gz -o $TGZ
     if [ `check_tgz $TGZ` != 1 ]; then
 	rm -f libxc-${VERSION}.tar.gz
 	curl -L https://github.com/ElectronicStructureLibrary/libxc/archive/refs/tags/${VERSION}.tar.gz -o $TGZ
@@ -78,7 +89,7 @@ fi
 
 cd libxc
 # patch pk09 to avoid compiler  memory problems
-patch -p0 -N < ../pk09.patch
+#patch -p0 -N < ../pk09.patch
 mkdir -p build
 cd build
 if [[ -z "${NWCHEM_TOP}" ]]; then
@@ -100,8 +111,15 @@ else
     fcflags=" "
 fi
 rm -rf libxc/build
+if [[ "${USE_HWOPT}" == "n" ]]; then
+    enable_xhost_flag=OFF
+else
+    enable_xhost_flag=ON
+fi
+
 $CMAKE -E env CFLAGS="$cflags" LDFLAGS="$ldflags" FCFLAGS="$fcflags" FFLAGS="$fcflags" \
 $CMAKE  -DCMAKE_INSTALL_PREFIX=${NWCHEM_TOP}/src/libext/libxc/install -DCMAKE_C_COMPILER=$CC -DENABLE_FORTRAN=ON -DCMAKE_Fortran_COMPILER=$FC -DDISABLE_KXC=OFF \
+-DENABLE_XHOST="$enable_xhost_flag" \
 -DCMAKE_INSTALL_LIBDIR="lib" -DCMAKE_BUILD_TYPE=Release ..
 
 make -j4 | tee make.log
